CHLRawGenerator 클래스 작성
